requirements.txt
openai-whisper==20231117
spacy==3.6.0
pandas==2.0.3